requirements.txt
numpy
scipy
pyaudio
pillow
pygame==2.5.0
pygame_gui==0.6.9
PyOpenGL
moviepy